pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Denys Chernousov <blackInOwls@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "0.6.0"
anyio = "4.3.0"
click = "8.1.7"
colorama = "0.4.6"
exceptiongroup = "1.2.0"
fastapi = "0.110.0"
h11 = "0.14.0"
httptools = "0.6.1"
idna = "3.6"
pydantic = "2.6.3"
pydantic-core = "2.16.3"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
sniffio = "1.3.1"
starlette = "0.36.3"
typing-extensions = "4.10.0"
uvicorn = "0.28.0"
watchfiles = "0.21.0"
websockets = "12.0"
orjson = "3.9.15"
sqlalchemy = "2.0.28"
alembic = "1.13.1"
asyncpg = "0.29.0"
fastapi-users = { version = "12.1.3", extras = ["sqlalchemy"] }
fastapi-cache2 = { version = "0.2.1", extras = ["redis"] }
celery = "^5.3.6"
flower = "^2.0.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
httpx = "^0.27.0"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [
    ".", "app"
]

